requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Deprecated==1.2.13
Django==4.0.6
django-extensions==3.2.0
djangorestframework==3.13.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
mssql-django==1.1.3
oauth2client==4.1.3
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydot==1.4.2
pydotplus==2.0.2
pyodbc==4.0.34
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.3
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
wrapt==1.14.1
